Update GRPC host address to environment variable
The hardcoded reference to the GRPC host address in the DdnsUpdater and SyncService classes has been replaced with an environment variable. This change allows for more flexibility and configurability, as the address can be easily changed without the need to modify the code. The corresponding environment variables have also been added to the docker-compose file.