Update Protobuf messages, alter URL and add environment variable
A new 'group' attribute was added to Protocol Buffers messages, which adjusts the ordering of existing fields. Some URLs in 'ProxmoxService.cs' have been changed. A GrpcChannel definition in 'DdnsUpdater.cs' was updated. A new environment variable "PROXMOX_FIREWALL_GROUP" was introduced in 'docker-compose.yml' and utilized in 'SyncService.cs'.