Change GrpcChannel to non-readonly and assign it
The `GrpcChannel` in `SyncService` was changed from a readonly variable to a non-readonly variable. This was done to assign it within the `ProcessPortRequest` method, allowing the GRPC channel to be appropriately set up and utilized by the `ProxmoxServiceClient`. BaseClient.dll to DdnsUpdater.dll in the final ENTRYPOINT has been applied.